Rename useEntityRecordBuilder to useEntityRecordCreate
